Move PeakCollection into the correct sub-package
Refs #219
