+++ pyrs/core/peak_fit_engine.py
@@ -247,7 +247,7 @@ class PeakFitEngine(object):
 
         Returns
         -------
-        List of ~pyrs.core.peak_collection.PeakCollection
+        List of ~pyrs.peaks.PeakCollection
 
         """
         raise NotImplementedError('Virtual base class member method fit_multiple_peaks')
@@ -346,7 +346,7 @@ class PeakFitEngine(object):
 
         Returns
         -------
-        pyrs.core.peak_collection.PeakCollection
+        pyrs.peaks.PeakCollection
             Collection of peak information
 
         """
