+++ pyrs/peaks/__init__.py
@@ -1,0 +1,3 @@
+from __future__ import (absolute_import, division, print_function)  # python3 compatibility
+
+from .peak_collection import PeakCollection
