+++ pyrs/utilities/rs_project_file.py
@@ -6,7 +6,7 @@ import numpy
 import os
 from pyrs.utilities import checkdatatypes
 from pyrs.core.instrument_geometry import AnglerCameraDetectorGeometry, HidraSetup
-from pyrs.core.peak_collection import PeakCollection
+from pyrs.peaks import PeakCollection
 from pyrs.dataobjects import SampleLogs
